// src/app/modules/tour/tour.interface.ts
export type TCreateTourPayload = {
    title: string;
    description: string;
    price: number;
    duration: number;
    meetingPoint: string;
    maxGroupSize: number;
    images: string[];
    location: string;
    isActive: boolean;
    itinerary: string;
    city: string;
    languages: string[];
    category: string


};